Proposal: "depth" enum conform to defacto standard
Per https://github.com/ocsf/ocsf-schema/issues/422, there was an action taken to number "enum" elements where 0 = Unknown and 99 = Other. This was not applied to depth ( with 0 = Unknown and 99 = Other.), resulting in the only instance of a non-numbered "enum". In the thread around issue 422 above, this was noted as a deliberate choice "usually used when the values are defined by external standard". However, this is the only instance in the dictionary where this appears to be the case. 

I am proposing that depth be changed to conform to style suggested by the other enums, retaining the textual caption. 

Signed-off-by: zhanif3 <zh4990@gmail.com>